Create DeleteClient.java

+++ Week-7/concurso/BackEmpresa/src/main/java/com/backempresa/Client/Infrastructure/Controller/V0/DeleteClient.java
@@ -1,6 +1,7 @@
 package com.backempresa.Client.Infrastructure.Controller.V0;
 
 import com.backempresa.Client.Application.ClientService;
+import com.backempresa.Utils.Exceptions.customUnprocesableException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,6 +18,9 @@ public class DeleteClient {
 
     @DeleteMapping("{id}")
     public void deleteTrip(@PathVariable UUID id) {
+        if (clientService.filterClientByEmail("admin@adminbus.local").getIdClient().equals(id)) {
+            throw new customUnprocesableException("No se puede eliminar administrador");
+        }
         clientService.deleteClient(id);
     }
 }
